+++ packages/react-server/src/client-internal.ts
@@ -1,4 +1,4 @@
 "use client";
 
-export { createServerReference } from "./lib/shared";
+export { createServerReference } from "./features/server-action/client";
 export { ErrorBoundary } from "./lib/client/error-boundary";
