+++ packages/react-server/src/entry/react-server.tsx
@@ -1,10 +1,11 @@
 import { createDebug, objectMapKeys } from "@hiogawa/utils";
 import reactServerDomServer from "react-server-dom-webpack/server.edge";
+import { ejectActionId } from "../features/server-action/utils";
+import { unwrapRscRequest } from "../features/server-component/utils";
 import { createBundlerConfig } from "../features/use-client/react-server";
 import { ReactServerDigestError, createError } from "../lib/error";
 import { __global } from "../lib/global";
 import { generateRouteTree, matchRoute, renderMatchRoute } from "../lib/router";
-import { ejectActionId, unwrapRscRequest } from "../lib/shared";
 
 const debug = createDebug("react-server:rsc");
