+++ packages/react-server/src/features/server-action/react-server.tsx
@@ -1,0 +1,18 @@
+export function createServerReference(id: string, action: Function): React.FC {
+  return Object.defineProperties(action, {
+    $$typeof: {
+      value: Symbol.for("react.server.reference"),
+    },
+    $$id: {
+      value: id,
+      configurable: true,
+    },
+    $$bound: { value: null, configurable: true },
+    bind: {
+      value: () => {
+        throw new Error("todo: createServerReference.bind");
+      },
+      configurable: true,
+    },
+  }) as any;
+}
