+++ packages/react-server/src/server-internal.ts
@@ -1,2 +1,2 @@
 export { createClientReference } from "./features/use-client/react-server";
-export { createServerReferenceForRsc as createServerReference } from "./lib/shared";
+export { createServerReference } from "./features/server-action/react-server";
